feat: NewDiaryViewController 삭제 및 DetailViewController 수정
