+++ Diary/View/Base.lproj/Main.storyboard
@@ -63,58 +63,6 @@
             </objects>
             <point key="canvasLocation" x="1057.2519083969464" y="-27.464788732394368"/>
         </scene>
-        <!--New Diary View Controller-->
-        <scene sceneID="tPl-Qg-dkH">
-            <objects>
-                <viewController storyboardIdentifier="NewDiaryViewController" id="uer-DR-yxf" customClass="NewDiaryViewController" customModule="Diary" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="9vH-tr-yIy">
-                        <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" spacing="21" translatesAutoresizingMaskIntoConstraints="NO" id="kXF-lU-8KL">
-                                <rect key="frame" x="16" y="109" width="361" height="689"/>
-                                <subviews>
-                                    <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="pON-7i-Xaq">
-                                        <rect key="frame" x="0.0" y="0.0" width="361" height="35"/>
-                                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                                        <constraints>
-                                            <constraint firstAttribute="height" constant="35" id="a53-nh-3JC"/>
-                                        </constraints>
-                                        <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
-                                        <color key="textColor" systemColor="labelColor"/>
-                                        <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                                        <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
-                                    </textView>
-                                    <textView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" textAlignment="natural" translatesAutoresizingMaskIntoConstraints="NO" id="cxe-2T-4Yj">
-                                        <rect key="frame" x="0.0" y="56" width="361" height="633"/>
-                                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                                        <string key="text">Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.</string>
-                                        <color key="textColor" systemColor="labelColor"/>
-                                        <fontDescription key="fontDescription" type="system" pointSize="14"/>
-                                        <textInputTraits key="textInputTraits" autocapitalizationType="sentences"/>
-                                    </textView>
-                                </subviews>
-                                <viewLayoutGuide key="safeArea" id="St8-0v-dEy"/>
-                            </stackView>
-                        </subviews>
-                        <viewLayoutGuide key="safeArea" id="EJd-ZS-Ods"/>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                        <constraints>
-                            <constraint firstItem="kXF-lU-8KL" firstAttribute="leading" secondItem="EJd-ZS-Ods" secondAttribute="leading" constant="16" id="aAU-zp-y1n"/>
-                            <constraint firstItem="kXF-lU-8KL" firstAttribute="top" secondItem="EJd-ZS-Ods" secondAttribute="top" constant="50" id="efI-FO-A6V"/>
-                            <constraint firstItem="EJd-ZS-Ods" firstAttribute="bottom" secondItem="kXF-lU-8KL" secondAttribute="bottom" constant="20" id="lHQ-V4-CrF"/>
-                            <constraint firstItem="EJd-ZS-Ods" firstAttribute="trailing" secondItem="kXF-lU-8KL" secondAttribute="trailing" constant="16" id="u6E-53-rrC"/>
-                        </constraints>
-                    </view>
-                    <connections>
-                        <outlet property="bodyTextView" destination="cxe-2T-4Yj" id="WJs-31-TDB"/>
-                        <outlet property="titleTextView" destination="pON-7i-Xaq" id="Tav-mh-WGe"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="qcl-ID-KIk" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="1878.6259541984732" y="-27.464788732394368"/>
-        </scene>
         <!--Detail View Controller-->
         <scene sceneID="y1i-RE-Htl">
             <objects>
@@ -165,7 +113,7 @@
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Qwl-AI-M76" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
             </objects>
-            <point key="canvasLocation" x="2661.8320610687024" y="-27.464788732394368"/>
+            <point key="canvasLocation" x="1755" y="-27"/>
         </scene>
         <!--Navigation Controller-->
         <scene sceneID="Jhy-bJ-j87">
